model/requirements_docker.txt
flask
tensorflow-cpu
pillow